Push for: compressed node storage, and using longs for node pointers.
git-svn-id: svn://svn.code.sf.net/p/meddly/code/branches/holes@414 14966d26-f10a-4722-9547-d5dcc201c352

+++ src/defines.h
@@ -87,7 +87,7 @@ namespace MEDDLY {
   template <class T> inline bool POSITIVE(T X) { return (X>0) ? true : false; }
 
   // Number of digits
-  inline int digits(int a) {
+  inline int digits(long a) {
     int d;
     for (d=1; a; d++) { a /= 10; }
     return d;
